+++ src/hooks/login/index.ts
@@ -1,34 +1,3 @@
-import { postKakaoLogin } from "@/apis/login";
-import { MutateOptions, useMutation } from "@tanstack/react-query";
+import useKakaoLogin from "@/hooks/login/useKakaoLogin";
 
-export const useKakaoLogin = () => {
-  const { postKakaoLogin } = useKakaoLoginMutate();
-
-  const onKakaoLogin = () => {
-    postKakaoLogin({
-      onError: () => {
-        alert("카카오 로그인에 실패했습니다.");
-      },
-    });
-  };
-  return {
-    onKakaoLogin,
-  };
-};
-
-export const useKakaoLoginMutate = () => {
-  const { mutate } = useMutation({
-    mutationKey: ["kakaoLogin"],
-    mutationFn: postKakaoLogin,
-  });
-
-  return {
-    postKakaoLogin: (
-      mutateOption?: MutateOptions<void, Error, void, unknown>
-    ) => {
-      mutate(undefined, {
-        ...mutateOption,
-      });
-    },
-  };
-};
+export { useKakaoLogin };
